Add 'article_id' parameter to 'getHeadlines' API method
This implements an 'article_id' parameter, analogously to
'getArticle'.  That is, it accepts either a single integer or a
comma-separated string of integers.

The user-visible behavior is:

- if both 'feed_id' and 'article_id' parameters are given, the results
are intersected

- if 'feed_id' is not given but 'article_id' is, 'feed_id' takes a
  default value of -4 (all feeds)

- if 'article_id' is not given, no changes are apparent

The implementation itself modifies the `api_get_headlines' method
signature, by appending a default argument '$article_id = false'.
Accordingly, it also modifies the 'queryFeedHeadlines' function
signature with the same change. This ensures the API remains backwards
compatible with other existing consumers.